# services/b2b/openapi.yaml
openapi: 3.0.3
info:
  title: NeoMarket B2B (Seller Cabinet)
  version: "1.1"
  description: |
    Кабинет продавца. Создание товаров, SKU, управление ценами, накладные.
    
    ## Статусы товара
    - `CREATED` — создан, ожидает отправки на модерацию
    - `ON_MODERATION` — на проверке у модератора
    - `MODERATED` — одобрен, виден покупателям
    - `BLOCKED` — заблокирован, продавец видит причину
servers:
  - url: http://localhost:8000
paths:
...
  /api/v1/skus:
    post:
      summary: Создать SKU
      tags: [SKU]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - product_id
                - name
                - price
              properties:
                product_id:
                  type: integer
                name:
                  type: string
                seller_sku:
                  type: string
                barcode:
                  type: string
                price:
                  type: integer
                  minimum: 0
                compare_at_price:
                  type: integer
                quantity:
                  type: integer
                  default: 0
                characteristics:
                  type: array
                  items:
                    $ref: "../shared/schemas.yaml#/components/schemas/CharacteristicValue"
      responses:
        "201":
          description: SKU создан
          content:
            application/json:
              schema:
                $ref: "../shared/schemas.yaml#/components/schemas/SKU"
    
    put:
      summary: Изменить SKU
      tags: [SKU]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                price:
                  type: integer
                compare_at_price:
                  type: integer
                is_active:
                  type: boolean
                characteristics:
                  type: array
                  items:
                    $ref: "../shared/schemas.yaml#/components/schemas/CharacteristicValue"
      responses:
        "200":
          description: SKU обновлён
          content:
            application/json:
              schema:
                $ref: "../shared/schemas.yaml#/components/schemas/SKU"

  # ========== НАКЛАДНЫЕ ==========
...